Hex command
The hex command may seem like another useless command  when in reality, it is very useful. It can let you test the different hex code colors to see how they will appear. This is especially useful when changing the CSS and you want to see how a color will look on PS. The hex command also allows you to find out the hex code of your own name (courtesy of Mr. Panpawn)! However, it doesn't allow you to find out the hex code of another users name due to privacy issues and because I didn't know how to check for a hex code and a username. But all in all, the hex command will be an awesome addition to the wide array of Frost commands and will also prove to be useful.