Rewrite /away and /back, "request" is not a module
Deleting the user object found creates a memory leak because the
reference to the deleted user object is kept after forcerenaming the
user. This command can now be used to add any kind of away status, so
long as it fits in the username.

"request" is no longer required, and instances where the "module" is
used are replaced with http.request.