/away style fixes, fix indentation
If the user is already away, it should check the length of the actual
username, not the username plus the status, when checking if the new
status makes the username too long. In the messages sent if the user's
driver or up, /away no longer includes the status, and /back puts the
status to lowercase.